requirement-prod.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.14
pymysql==1.1.0
cryptography==41.0.7
python-multipart==0.0.6
pydantic-settings==2.1.0
passlib==1.7.4
python-jose==3.3.0
gunicorn==21.2.0
python-dotenv==1.0.0